removes redundant code and refactors package structure
