+++ daraja/src/main/java/com/androidstudy/daraja/callback/DarajaPaymentListener.kt
@@ -15,7 +15,7 @@
  */
 package com.androidstudy.daraja.callback
 
-import com.androidstudy.daraja.model.PaymentResult
+import com.androidstudy.daraja.data.model.PaymentResult
 
 interface DarajaPaymentListener {
     fun onPaymentRequestComplete(result: PaymentResult)
