+++ daraja/src/main/java/com/androidstudy/daraja/data/model/AccessToken.kt
@@ -13,6 +13,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.androidstudy.daraja.model
+package com.androidstudy.daraja.data.model
 
 class AccessToken(var access_token: String, var expiresIn: String)
