+++ daraja/src/main/java/com/androidstudy/daraja/data/model/PaymentResult.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.androidstudy.daraja.model
+package com.androidstudy.daraja.data.model
 
 data class PaymentResult(
     var MerchantRequestID: String,
