+++ daraja/src/main/java/com/androidstudy/daraja/network/AuthAPI.kt
@@ -15,7 +15,7 @@
  */
 package com.androidstudy.daraja.network
 
-import com.androidstudy.daraja.model.AccessToken
+import com.androidstudy.daraja.data.model.AccessToken
 import retrofit2.Call
 import retrofit2.http.GET
